requirements.txt
audible>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1
rich>=13.7.0
cryptography>=41.0.0 